% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_grid.R
\name{evapo_l8t_grid}
\alias{evapo_l8t_grid}
\title{Actual evapotranspiration (ETa) using Landsat-8 (including thermal bands) images with a grid of agrometeorological data.}
\usage{
evapo_l8t_grid(doy, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24"), NDVI, Surface Temperature ("LST"), net radiation ("Rn_MJ"), Crop Coefficient ("kc") and Actual Evapotranspiration (evapo).
}
\description{
Actual evapotranspiration (ETa) using Landsat-8 (including thermal bands) images with a grid of agrometeorological data.
}
