% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_single.R
\name{evapo_modis}
\alias{evapo_modis}
\title{Actual evapotranspiration (ETa) using MODIS with single agrometeorological data.}
\usage{
evapo_modis(doy, RG, Ta, ET0, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{RG}{is the global solar radiation}

\item{Ta}{is the average air temperature}

\item{ET0}{is the reference evapotranspiration}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24"), NDVI, Surface Temperature ("LST"), net radiation ("Rn_MJ"), Crop Coefficient ("kc") and Actual Evapotranspiration (evapo).
}
\description{
Actual evapotranspiration (ETa) using MODIS with single agrometeorological data.
}
\examples{
library(agriwater)

# dependencies of package 'agriwater'
library(sp)
library(raster)
library(rgdal)

# creating raster which simulate Sentinel-2 reflectances - for using
# real data, please download:
# https://drive.google.com/open?id=14E1wHNLxG7_Dh4I-GqNYakj8YJDgKLzk

xy <- matrix(rnorm(100, mean = 0.07, sd = 0.01), 10, 10)
rast <- raster(xy)
extent(rast) <- c(-40.5,-40.45,-9.5,-9.45)
projection(rast) <- CRS("+proj=longlat +datum=WGS84")
writeRaster(rast, "B2", format = "GTiff", overwrite=TRUE)
xy <- matrix(rnorm(100, mean = 0.05, sd = 0.015),10, 10)
rast <- raster(xy)
extent(rast) <- c(-40.5,-40.45,-9.5,-9.45)
projection(rast) <- CRS("+proj=longlat +datum=WGS84")
writeRaster(rast, "B1", format = "GTiff", overwrite=TRUE)

mask <- as(extent(rast), 'SpatialPolygons')
projection(mask) <- CRS("+proj=longlat +datum=WGS84")
shapefile(mask, "mask.shp", overwrite=TRUE)

# using "agriwater" - it's the same procedure as the used for
# evapo_l8(), evapo_l8t(), evapo_modis_grid(), evapo_l8_grid(),
# evapo_l8t_grid(), evapo_s2() and evapo_s2_grid()
DOY = 134
a = 1.8
b = -0.008
RG = 17.6
Ta = 27.9
ET0 = 3.8
evapo_modis(DOY, RG, Ta, ET0, a, b)
}
