% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_grid.R, R/landsat8_single.R
\name{kc_l8}
\alias{kc_l8}
\title{Crop coefficient (ETa / ET0) using Landsat-8 images with a grid of agrometeorological data.}
\usage{
kc_l8(doy, RG, Ta, a, b)

kc_l8(doy, RG, Ta, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{RG}{is the global solar radiation}

\item{Ta}{is the average air temperature}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}

\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\description{
Crop coefficient (ETa / ET0) using Landsat-8 images with a grid of agrometeorological data.

Crop coefficient (ETa / ET0) using Landsat-8 images with single agrometeorological data.
}
