% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_single.R
\name{kc_l8t}
\alias{kc_l8t}
\title{Crop coefficient (ETa / ET0) using Landsat-8 images (including thermal bands) with single agrometeorological data.}
\usage{
kc_l8t(doy, RG, Ta, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{RG}{is the global solar radiation}

\item{Ta}{is the average air temperature}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24"), NDVI, Surface Temperature ("LST"), Crop Coefficient ("kc") and net radiation ("Rn_MJ").
}
\description{
Crop coefficient (ETa / ET0) using Landsat-8 images (including thermal bands) with single agrometeorological data.
}
