% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_grid.R
\name{kc_modis_grid}
\alias{kc_modis_grid}
\title{Crop coefficient (ETa / ET0) using MODIS with a grid of agrometeorological data.}
\usage{
kc_modis_grid(doy, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24"), NDVI, Surface Temperature ("LST"), Crop Coefficient ("kc") and net radiation ("Rn_MJ").
}
\description{
Crop coefficient (ETa / ET0) using MODIS with a grid of agrometeorological data.
}
