% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_grid.R
\name{radiation_l8t_grid}
\alias{radiation_l8t_grid}
\title{Energy balance using Landsat-8 images (including thermal bands) with a grid of agrometeorological data.}
\usage{
radiation_l8t_grid(dj, a, b)
}
\arguments{
\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}

\item{doy}{is the Day of Year (DOY)}
}
\description{
Energy balance using Landsat-8 images (including thermal bands) with a grid of agrometeorological data.
}
