% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2_grid.R
\name{radiation_s2_grid}
\alias{radiation_s2_grid}
\title{Energy balance using Sentinel-2 images with a grid of agrometeorological data.}
\usage{
radiation_s2_grid(doy, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24"), NDVI, Surface Temperature ("LST"), Crop Coefficient ("kc"), Actual Evapotranspiration (evapo), latent heat flux "LE_MJ"), net radiation ("Rn_MJ"), ground heat flux ("G_MJ") and the sensible heat flux ("H_MJ").
}
\description{
Energy balance using Sentinel-2 images with a grid of agrometeorological data.
}
