% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_single.R
\name{reflectance_l8}
\alias{reflectance_l8}
\title{Reflectancies from Landsat-8 images.}
\usage{
reflectance_l8(doy)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24").
}
\description{
Reflectancies from Landsat-8 images.
}
