% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_single.R
\name{albedo_l8}
\alias{albedo_l8}
\title{Surface Albedo using Landsat-8 images.}
\usage{
albedo_l8(doy)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24").
}
\description{
Surface Albedo using Landsat-8 images.
}
