% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_single.R
\name{albedo_modis}
\alias{albedo_modis}
\title{Surface Albedo using MODIS images.}
\usage{
albedo_modis()
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24").
}
\description{
Surface Albedo using MODIS images.
}
