% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2_single.R
\name{albedo_s2}
\alias{albedo_s2}
\title{Surface Albedo using Sentinel-2 images.}
\usage{
albedo_s2()
}
\value{
It returns in raster format (.tif) the Surface Albedo at 24h scale ("Alb_24").
}
\description{
Surface Albedo using Sentinel-2 images.
}
