% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_single.R
\name{evapo_l8t}
\alias{evapo_l8t}
\title{Actual evapotranspiration (ETa) using Landsat-8 (including thermal bands) images with single agrometeorological data.}
\usage{
evapo_l8t(doy, RG, ET0, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{RG}{is the global solar radiation}

\item{ET0}{is the reference evapotranspiration}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\description{
Actual evapotranspiration (ETa) using Landsat-8 (including thermal bands) images with single agrometeorological data.
}
