% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modis_grid.R
\name{evapo_modis_grid}
\alias{evapo_modis_grid}
\title{Actual evapotranspiration (ETa) using MODIS with a grid of agrometeorological data.}
\usage{
evapo_modis_grid(doy, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\description{
Actual evapotranspiration (ETa) using MODIS with a grid of agrometeorological data.
}
