% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_single.R
\name{kc_l8t}
\alias{kc_l8t}
\title{Crop coefficient (ETa / ET0) using Landsat-8 images (including thermal bands) with single agrometeorological data.}
\usage{
kc_l8t(doy, RG, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{RG}{is the global solar radiation}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\description{
Crop coefficient (ETa / ET0) using Landsat-8 images (including thermal bands) with single agrometeorological data.
}
