% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landsat8_thermal_single.R
\name{radiation_l8t}
\alias{radiation_l8t}
\title{Energy balance using Landsat-8 images (including thermal bands) with single agrometeorological data.}
\usage{
radiation_l8t(dj, RG, ET0, a, b)
}
\arguments{
\item{RG}{is the global solar radiation}

\item{ET0}{is the reference evapotranspiration}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}

\item{doy}{is the Day of Year (DOY)}
}
\description{
Energy balance using Landsat-8 images (including thermal bands) with single agrometeorological data.
}
