% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel2_grid.R
\name{radiation_s2_grid}
\alias{radiation_s2_grid}
\title{Energy balance using Sentinel-2 images with a grid of agrometeorological data.}
\usage{
radiation_s2_grid(doy, a, b)
}
\arguments{
\item{doy}{is the Day of Year (DOY)}

\item{a}{is one of the regression coefficients of SAFER algorithm}

\item{b}{is one of the regression coefficients of SAFER algorithm}
}
\description{
Energy balance using Sentinel-2 images with a grid of agrometeorological data.
}
